// src/entities/Product/model/types/product.ts
export interface Product {
    id: number;
    title: string;
    price: number;
    description?: string;
    category?: string;
    image?: string;
    rating?: {
      rate: number;
      count: number;
    };
}

export interface ProductSchema {
    product?: Product;
    isLoading: boolean;
    error?: string;
}